moved serialization files out of text into separate module
